requirements.txt
Flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.11.1
psutil==5.9.8
requests==2.31.0
python-whois==0.8.0
pywebview==5.3
